+++ src/main/java/com/fitness/service/impl/CourseInfoServiceImpl.java
@@ -1,6 +1,7 @@
 package com.fitness.service.impl;
 
 import com.fitness.dao.CourseInfoMapper;
+import com.fitness.dao.CourseMapper;
 import com.fitness.entity.CourseInfo;
 import com.fitness.entity.Result;
 import com.fitness.exception.PageNumberException;
@@ -25,9 +26,12 @@ public class CourseInfoServiceImpl implements CourseInfoService {
 
   private CourseInfoMapper courseInfoMapper;
 
+  private CourseMapper courseMapper;
+
   @Autowired
-  public CourseInfoServiceImpl(CourseInfoMapper courseInfoMapper) {
+  public CourseInfoServiceImpl(CourseInfoMapper courseInfoMapper, CourseMapper courseMapper) {
     this.courseInfoMapper = courseInfoMapper;
+    this.courseMapper = courseMapper;
   }
   @Override
   public Result<PageInfo<CourseInfo>> getCourseInfoByPage(Integer startPage, Integer pageSize) {
@@ -69,6 +73,9 @@ public class CourseInfoServiceImpl implements CourseInfoService {
 
   @Override
   public Result deleteCourseInfo(String id) {
+    if(courseMapper.getCourseInfoNumber(id) > 0) {
+      return new Result(Result.OK, "该课程被选修中，暂不能删除~");
+    }
     courseInfoMapper.deleteCourseInfo(id);
     return new Result(Result.OK, "删除成功~");
   }
